remove alpha and adjust PR
