+++ pkg/controllers/disruption/validation.go
@@ -77,7 +77,7 @@ func (v *Validation) IsValid(ctx context.Context, cmd Command) (bool, error) {
 		case <-v.clock.After(waitDuration):
 		}
 	}
-	validationCandidates, err := GetCandidates(ctx, v.cluster, v.kubeClient, v.recorder, v.clock, v.cloudProvider, v.ShouldDeprovision, v.queue)
+	validationCandidates, err := GetCandidates(ctx, v.cluster, v.kubeClient, v.recorder, v.clock, v.cloudProvider, v.ShouldDisrupt, v.queue)
 	if err != nil {
 		return false, fmt.Errorf("constructing validation candidates, %w", err)
 	}
@@ -115,11 +115,8 @@ func (v *Validation) ShouldDisrupt(_ context.Context, c *Candidate) bool {
 	return c.nodePool.Spec.Disruption.ConsolidationPolicy == v1beta1.ConsolidationPolicyWhenUnderutilized
 }
 
-// ValidateCommand validates a command for a deprovisioner
+// ValidateCommand validates a command for a Method
 func (v *Validation) ValidateCommand(ctx context.Context, cmd Command, candidates []*Candidate) (bool, error) {
-	// map from candidates we are about to remove back into candidates with cluster state
-	candidates = mapCandidates(cmd.candidates, candidates)
-
 	// None of the chosen candidate are valid for execution, so retry
 	if len(candidates) == 0 {
 		return false, nil
