+++ pkg/test/expectations/expectations.go
@@ -582,49 +582,3 @@ func ExpectMakeNodesAndNodeClaimsInitializedAndStateUpdated(ctx context.Context,
 		ExpectReconcileSucceeded(ctx, nodeClaimStateController, client.ObjectKeyFromObject(m))
 	}
 }
-
-func ExpectMakeNewNodeClaimsReady(ctx context.Context, c client.Client, wg *sync.WaitGroup, cluster *state.Cluster,
-	cloudProvider cloudprovider.CloudProvider, numNewNodeClaims int) {
-	GinkgoHelper()
-
-	existingNodeClaims := ExpectNodeClaims(ctx, c)
-	existingNodeClaimNames := sets.NewString(lo.Map(existingNodeClaims, func(nc *v1beta1.NodeClaim, _ int) string {
-		return nc.Name
-	})...)
-
-	wg.Add(1)
-	go func() {
-		nodeClaimsMadeReady := 0
-		ctx, cancel := context.WithTimeout(ctx, time.Second*10) // give up after 10s
-		defer GinkgoRecover()
-		defer wg.Done()
-		defer cancel()
-		for {
-			select {
-			case <-time.After(50 * time.Millisecond):
-				nodeClaimList := &v1beta1.NodeClaimList{}
-				if err := c.List(ctx, nodeClaimList); err != nil {
-					continue
-				}
-				for i := range nodeClaimList.Items {
-					nc := &nodeClaimList.Items[i]
-					if existingNodeClaimNames.Has(nc.Name) {
-						continue
-					}
-					nc, n := ExpectNodeClaimDeployed(ctx, c, cluster, cloudProvider, nc)
-					ExpectMakeNodeClaimsInitialized(ctx, c, nc)
-					ExpectMakeNodesInitialized(ctx, c, n)
-
-					nodeClaimsMadeReady++
-					existingNodeClaimNames.Insert(nc.Name)
-					// did we make all the nodes ready that we expected?
-					if nodeClaimsMadeReady == numNewNodeClaims {
-						return
-					}
-				}
-			case <-ctx.Done():
-				Fail(fmt.Sprintf("waiting for nodeclaims to be ready, %s", ctx.Err()))
-			}
-		}
-	}()
-}
